fix more bugs
- some of which were only introduced in the previous commits

+++ 1.16_combat-6/src/main/java/io/github/axolotlclient/modules/hud/gui/hud/KeystrokeHud.java
@@ -302,7 +302,7 @@ public class KeystrokeHud extends TextHudEntry {
 		protected final Rectangle bounds;
 		private final int animTime = 100;
 		protected DrawPosition offset;
-		private float start = -1;
+		private long start = -1;
 		private boolean wasPressed = false;
 
 		public Keystroke(Rectangle bounds, DrawPosition offset, KeyBinding key, KeystrokeRenderer render) {
@@ -318,7 +318,7 @@ public class KeystrokeHud extends TextHudEntry {
 		}
 
 		private float getPercentPressed() {
-			return start == -1 ? 1 : MathHelper.clamp((Util.getMeasuringTimeMs() - start) / animTime, 0, 1);
+			return start == -1 ? 1 : MathHelper.clamp((float)(Util.getMeasuringTimeMs() - start) / animTime, 0, 1);
 		}
 
 		public void render(MatrixStack matrices) {
@@ -337,7 +337,7 @@ public class KeystrokeHud extends TextHudEntry {
 			if (outline.get()) {
 				outlineRect(matrices, rect, getOutlineColor());
 			}
-			if ((Util.getMeasuringTimeMs() - start) / animTime >= 1) {
+			if ((float)(Util.getMeasuringTimeMs() - start) / animTime >= 1) {
 				start = -1;
 			}
 			wasPressed = key.isPressed();
